Increase the time for checking dockers log output
After the pr https://github.com/os-autoinst/os-autoinst-distri-openQA/pull/62
was merged, sometimes the jobs failed again. Because the first time getting logs
doesn't include the content that we need to check. So we should try more
times to get it.
Here is a failed example: https://openqa.opensuse.org/tests/1670864#